Gilad And Dennis Projects/Backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5
chardet==3.0.4
colorama==0.4.3
Django==3.0.4
djangorestframework==3.11.0
httpie==2.0.0
idna==2.9
mysqlclient==1.4.6
Pygments==2.6.1
pytz==2019.3
requests==2.23.0
splunk-sdk==1.6.12
sqlparse==0.3.1
urllib3==1.25.8
